+++ admin/sql.inc.php
@@ -64,6 +64,7 @@ if ($_POST) {
 
 		$space = "(?:\\s|/\\*[\s\S]*?\\*/|(?:#|$line_comment)[^\n]*\n?|--\r?\n)";
 		$delimiter = ";";
+		$delimiter_length = 1;
 		$offset = 0;
 		$empty = true;
 
@@ -85,7 +86,8 @@ if ($_POST) {
 
 		while ($query != "") {
 			if (!$offset && preg_match("~^$space*+DELIMITER\\s+(\\S+)~i", $query, $match)) {
-				$delimiter = $match[1];
+				$delimiter = preg_quote($match[1]);
+				$delimiter_length = strlen($match[1]);
 
 				$formatted_query = Admin::get()->formatSqlCommandQuery(trim($match[0]));
 				if ($formatted_query != "") {
@@ -93,8 +95,12 @@ if ($_POST) {
 				}
 
 				$query = substr($query, strlen($match[0]));
+			} elseif (!$offset && DIALECT == "pgsql" && preg_match("~^($space*+COPY\\s+)[^;]+\\s+FROM\\s+stdin;~i", $query, $match)) {
+				$delimiter = "\n\\\\\\.\r?\n";
+				$delimiter_length = 3;
+				$offset = strlen($match[0]);
 			} else {
-				preg_match('(' . preg_quote($delimiter) . "\\s*|$parse)", $query, $match, PREG_OFFSET_CAPTURE, $offset); // should always match
+				preg_match("($delimiter\\s*|$parse)", $query, $match, PREG_OFFSET_CAPTURE, $offset); // always matches
 				/** @var int $pos */
 				list($found, $pos) = $match[0];
 				if (!$found && $fp && !feof($fp)) {
@@ -105,7 +111,7 @@ if ($_POST) {
 					}
 					$offset = $pos + strlen($found);
 
-					if ($found && rtrim($found) != $delimiter) { // find matching quote or comment end
+					if ($found && !preg_match("(^$delimiter)", $found)) { // find matching quote or comment end
 						$c_style_escapes = Driver::get()->hasCStyleEscapes() || (DIALECT == "pgsql" && ($pos > 0 && strtolower($query[$pos - 1]) == "e"));
 
 						$pattern = '(';
@@ -134,7 +140,7 @@ if ($_POST) {
 
 					} else { // end of a query
 						$empty = false;
-						$q = substr($query, 0, $pos + strlen($delimiter));
+						$q = substr($query, 0, $pos + $delimiter_length);
 						$commands++;
 						$print = "<pre id='sql-$commands'><code class='jush-" . DIALECT . "'>" . Admin::get()->formatSqlCommandQuery(trim($q)) . "</code></pre>\n";
 						if (DIALECT == "sqlite" && preg_match("~^$space*+ATTACH\\b~i", $q, $match)) {